importer/requirements.txt
flake8==3.0.4
docopt==0.6.2
PyYAML==3.12
elasticsearch==2.3.0
requests==2.11.0
xlrd==1.0.0
